pyproject.toml
[tool.poetry]
name = "boilerplate"
version = "0.1.0"
description = ""
authors = ["Krish Maharjan <krish.maharjan.work@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.6"
djangorestframework = "^3.15.1"
pika = "^1.3.2"
celery = "^5.4.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
